+++ Software Project/src/LandlineReceipt.java
@@ -2,7 +2,7 @@
 public interface LandlineReceipt {
 	public void landLineReceiptForm();
 
-	public int getAmount();
+	public double getAmount();
 	
 
 }
